JankenGameの中身をMainに移す
